cleaning up flutter lib
