+++ lib/products/enhancedselfie/smile_id_smart_selfie_authentication_enhanced.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/widgets.dart';
+
+import '../../views/smile_view.dart';
+
+class SmileIDSmartSelfieAuthenticationEnhanced extends StatelessWidget {
+  static const String viewType = "SmileIDSmartSelfieAuthenticationEnhanced";
+  final Map<String, dynamic> creationParams;
+
+  /// Called when the user successfully completes the smart selfie enrollment flow. The result is a
+  /// JSON string.
+  final Function(String) onSuccess;
+  final Function(String) onError;
+
+  const SmileIDSmartSelfieAuthenticationEnhanced._({
+    required this.creationParams,
+    required this.onSuccess,
+    required this.onError,
+  });
+
+  factory SmileIDSmartSelfieAuthenticationEnhanced({
+    Key? key,
+    // userId can't actually be null in the native SDK but we delegate their creation to
+    // the native platform code, since that's where the random ID creation happens
+    String? userId,
+    bool allowNewEnroll = false,
+    bool showAttribution = true,
+    bool showInstructions = true,
+    bool skipApiSubmission = false,
+    Map<String, String>? extraPartnerParams,
+    required Function(String resultJson) onSuccess,
+    required Function(String errorMessage) onError,
+  }) {
+    return SmileIDSmartSelfieAuthenticationEnhanced._(
+      onSuccess: onSuccess,
+      onError: onError,
+      creationParams: {
+        "userId": userId,
+        "allowNewEnroll": allowNewEnroll,
+        "showAttribution": showAttribution,
+        "showInstructions": showInstructions,
+        "skipApiSubmission": skipApiSubmission,
+        "extraPartnerParams": extraPartnerParams,
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return buildSmileIDPlatformView(
+      context: context,
+      viewType: viewType,
+      creationParams: creationParams,
+      onSuccess: onSuccess,
+      onError: onError,
+    );
+  }
+}
