--- lib/smile_id_document_capture_view.dart
@@ -1,107 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
-
-class SmileIDDocumentCaptureView extends StatelessWidget {
-  static const String viewType = "SmileIDDocumentCaptureView";
-  final Map<String, dynamic> creationParams;
-
-  /// Called when the user successfully completes the smart selfie enrollment flow. The result is a
-  /// JSON string.
-  final Function(String) onSuccess;
-  final Function(String) onError;
-
-  const SmileIDDocumentCaptureView._({
-    required this.creationParams,
-    required this.onSuccess,
-    required this.onError,
-  });
-
-  factory SmileIDDocumentCaptureView({
-    Key? key,
-    bool isDocumentFrontSide = true,
-    bool showInstructions = true,
-    bool showAttribution = true,
-    bool allowGalleryUpload = true,
-    bool showConfirmationDialog = true,
-    double? idAspectRatio,
-    required Function(String resultJson) onSuccess,
-    required Function(String errorMessage) onError,
-  }) {
-    return SmileIDDocumentCaptureView._(
-      onSuccess: onSuccess,
-      onError: onError,
-      creationParams: {
-        "isDocumentFrontSide": isDocumentFrontSide,
-        "showInstructions": showInstructions,
-        "showAttribution": showAttribution,
-        "allowGalleryUpload": allowGalleryUpload,
-        "showConfirmationDialog": showConfirmationDialog,
-        "idAspectRatio": idAspectRatio,
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return SafeArea(
-          child: PlatformViewLink(
-          viewType: viewType,
-          surfaceFactory: (context, controller) {
-            return AndroidViewSurface(
-              controller: controller as AndroidViewController,
-              hitTestBehavior: PlatformViewHitTestBehavior.opaque,
-              gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{
-                Factory<OneSequenceGestureRecognizer>(EagerGestureRecognizer.new)
-              },
-            );
-          },
-          onCreatePlatformView: (params) {
-            return PlatformViewsService.initExpensiveAndroidView(
-              id: params.id,
-              viewType: viewType,
-              layoutDirection: Directionality.of(context),
-              creationParams: creationParams,
-              creationParamsCodec: const StandardMessageCodec(),
-              onFocus: () {
-                params.onFocusChanged(true);
-              },
-            )
-              ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
-              ..addOnPlatformViewCreatedListener(_onPlatformViewCreated)
-              ..create();
-            },
-          ),
-        );
-      case TargetPlatform.iOS:
-        return UiKitView(
-          viewType: viewType,
-          creationParams: creationParams,
-          creationParamsCodec: const StandardMessageCodec(),
-          onPlatformViewCreated: _onPlatformViewCreated,
-        );
-      default:
-        throw UnsupportedError("Unsupported platform");
-    }
-  }
-
-  void _onPlatformViewCreated(int id) {
-    final channel = MethodChannel("${viewType}_$id");
-    channel.setMethodCallHandler(_handleMethodCall);
-  }
-
-  Future<dynamic> _handleMethodCall(MethodCall call) async {
-    switch (call.method) {
-      case "onSuccess":
-        onSuccess(call.arguments);
-      case "onError":
-        onError(call.arguments);
-      default:
-        throw MissingPluginException();
-    }
-  }
-}
